+++ src/main/java/com/bank/dao/CurrencyDAO.java
@@ -1,6 +1,8 @@
 package com.bank.dao;
 
+import java.math.BigDecimal;
+
 public interface CurrencyDAO {
 	
-	public double getCurrencyValue(String currencyName);
+	public BigDecimal getCurrencyValue(String currencyName);
 }
