+++ src/main/java/com/bank/service/TransferService.java
@@ -1,11 +1,13 @@
 package com.bank.service;
 
+import java.math.BigDecimal;
+
 public interface TransferService {
 
-	public void transfer(String senderAccountName, String recipientAccountName, double amount);
+	public void transfer(String senderAccountName, String recipientAccountName, BigDecimal amount);
 	
-	public void deposit(String accountName, double amount);
+	public void deposit(String accountName, BigDecimal amount);
 	
-	public void withdraw(String accountName, double amount);
+	public void withdraw(String accountName, BigDecimal amount);
 
 }
